Enabled complete stack
