--- manufacturing-ai-suite/industrial-edge-insights-multimodal/time_series_analytics_microservice/tick_scripts/windturbine_anomaly_detector.tick
@@ -1,22 +1,0 @@
-dbrp "datain"."autogen"
-
-var data0 = stream
-        |from()
-                .database('datain')
-                .retentionPolicy('autogen')
-                .measurement('wind_turbine_data')
-        @windturbine_anomaly_detector()
-data0
-        |influxDBOut()
-                .buffer(0)
-                .database('datain')
-                .measurement('wind_turbine_anomaly_data')
-                .retentionPolicy('autogen')
-data0
-        |alert()
-                .crit(lambda: "anomaly_status" > 0)
-                .message('Anomaly detected for wind speed: {{ index .Fields "wind_speed" }} Grid Active Power: {{ index .Fields "grid_active_power" }} Anomaly Status: {{ index .Fields "anomaly_status" }} ')
-                .noRecoveries()
-                .mqtt('my_mqtt_broker')
-                .topic('alerts/wind_turbine')
-                .qos(1)
